refactor: Add base and runner dir
