+++ packages/bridge/src/base/config.ts
